[TRIBE-120] refactor: touch event 추가 및 훅 추가해서 코드 리팩터링
