+++ src/components/control/index.ts
@@ -1,1 +1,1 @@
-export { default as Slider } from "./Slider";
+export { default as Slider } from "./slider/Slider";
